pyproject.toml
[tool.poetry]
name = "PaperAgent"
version = "0.1.0"
description = "An AI-powered assistant that lowers the bar-to-entry into specialized niches by providing an elaborate literature review."
authors = ["Jinal.Shah <28205508+JinalShah2002@users.noreply.github.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
arxiv = "^2.1.3"
anthropic = "^0.30.1"
python-dotenv = "^1.0.1"
PyPDF2 = "^3.0.1"
langchain = "^0.2.6"
pinecone-client = ">=3.2.2,<4.0.0"
openai = "^1.35.9"
langchain-openai = "^0.1.14"
langchain-anthropic = "^0.1.19"
langchain-core = "^0.2.11"
langchain-text-splitters = "^0.2.2"
langchain-pinecone = "^0.1.1"
torch = "2.1.2"
keybert = "^0.8.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
